+++ sdk/storage/azure_storage_blob/src/models/method_options.rs
@@ -10,8 +10,10 @@ use azure_core::{
 use time::OffsetDateTime;
 
 use crate::models::{
-    AccessTier, BlobClientDownloadInternalOptions, BlobContainerClientListBlobsInternalOptions,
-    EncryptionAlgorithmType, HttpRange, ImmutabilityPolicyMode, ListBlobsIncludeItem,
+    AccessTier, BlobClientDownloadInternalOptions,
+    BlobContainerClientListBlobsHierarchicalInternalOptions,
+    BlobContainerClientListBlobsInternalOptions, EncryptionAlgorithmType, HttpRange,
+    ImmutabilityPolicyMode, ListBlobsIncludeItem,
 };
 
 /// Options to be passed to `BlobClient::download()`
@@ -135,6 +137,9 @@ pub struct BlobContainerClientListBlobsOptions<'a> {
     /// Selects the response format. Defaults to [`ListBlobsAcceptFormat::Arrow`].
     pub accept: Option<ListBlobsAcceptFormat>,
 
+    /// Filters the results to return only names that are ordered before this value. Only applies to the Apache Arrow scenario.
+    pub end_before: Option<String>,
+
     /// Specify to include additional, optional information.
     pub include: Option<Vec<ListBlobsIncludeItem>>,
 
@@ -161,6 +166,7 @@ impl BlobContainerClientListBlobsOptions<'_> {
     pub(crate) fn into_owned(self) -> BlobContainerClientListBlobsOptions<'static> {
         BlobContainerClientListBlobsOptions {
             accept: self.accept,
+            end_before: self.end_before,
             include: self.include,
             marker: self.marker,
             maxresults: self.maxresults,
@@ -179,7 +185,73 @@ impl BlobContainerClientListBlobsOptions<'_> {
         method_options: ClientMethodOptions<'static>,
     ) -> BlobContainerClientListBlobsInternalOptions<'static> {
         BlobContainerClientListBlobsInternalOptions {
-            end_before: None,
+            end_before: self.end_before.clone(),
+            include: self.include.clone(),
+            marker: self.marker.clone(),
+            maxresults: self.maxresults,
+            method_options,
+            prefix: self.prefix.clone(),
+            start_from: self.start_from.clone(),
+            timeout: self.timeout,
+        }
+    }
+}
+
+/// Options to be passed to [`BlobContainerClient::list_blobs_hierarchical`](crate::BlobContainerClient::list_blobs_hierarchical).
+#[derive(Clone, Default, SafeDebug)]
+pub struct BlobContainerClientListBlobsHierarchicalOptions<'a> {
+    /// Selects the response format. Defaults to [`ListBlobsAcceptFormat::Arrow`].
+    pub accept: Option<ListBlobsAcceptFormat>,
+
+    /// Filters the results to return only names that are ordered before this value. Only applies to the Apache Arrow scenario.
+    pub end_before: Option<String>,
+
+    /// Specify to include additional, optional information.
+    pub include: Option<Vec<ListBlobsIncludeItem>>,
+
+    /// An opaque string value that identifies the portion of the result set to return with this operation.
+    pub marker: Option<String>,
+
+    /// Specifies the maximum number of resources to return.
+    pub maxresults: Option<i32>,
+
+    /// Allows customization of the method call.
+    pub method_options: PagerOptions<'a>,
+
+    /// Filters the results to return only resources whose name begins with the specified prefix.
+    pub prefix: Option<String>,
+
+    /// Specifies the relative path to list paths from.
+    pub start_from: Option<String>,
+
+    /// The timeout parameter is expressed in seconds.
+    pub timeout: Option<i32>,
+}
+
+impl BlobContainerClientListBlobsHierarchicalOptions<'_> {
+    pub(crate) fn into_owned(self) -> BlobContainerClientListBlobsHierarchicalOptions<'static> {
+        BlobContainerClientListBlobsHierarchicalOptions {
+            accept: self.accept,
+            end_before: self.end_before,
+            include: self.include,
+            marker: self.marker,
+            maxresults: self.maxresults,
+            method_options: PagerOptions {
+                context: self.method_options.context.into_owned(),
+                ..self.method_options
+            },
+            prefix: self.prefix,
+            start_from: self.start_from,
+            timeout: self.timeout,
+        }
+    }
+
+    pub(crate) fn to_internal(
+        &self,
+        method_options: ClientMethodOptions<'static>,
+    ) -> BlobContainerClientListBlobsHierarchicalInternalOptions<'static> {
+        BlobContainerClientListBlobsHierarchicalInternalOptions {
+            end_before: self.end_before.clone(),
             include: self.include.clone(),
             marker: self.marker.clone(),
             maxresults: self.maxresults,